fashion_store/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
Pillow>=10.3
python-dotenv>=1.0